Refactor out ModelSettings

+++ DronePan/ViewController.h
@@ -18,7 +18,7 @@
 @interface ViewController : UIViewController
 
 -(double) headingTo360:(double)heading;
-- (NSArray *)yawAnglesForCount:(int)count withHeading:(double)heading;
+- (NSArray *)yawAnglesForCount:(long)count withHeading:(double)heading;
 
 @end
